rename cmd.py to cmdline.py because 'cmd' exists in python standard library
