+++ scheme/eval.py
@@ -5,7 +5,7 @@ from prims import add_globals
 from errors import EvalError
 
 # for test only
-from cmd import Echo 
+from cmdline import Echo
 
 isa = isinstance
 
@@ -33,7 +33,7 @@ class Env(dict):
 
     def to_pretty_str(self, depth=0):
         """
-        Recursively generate the environments. 
+        Recursively generate the environments.
         Indented according to its depth.
         """
 
@@ -52,14 +52,14 @@ class Env(dict):
     #         for k in self.keys():
     #             envStr += ", {0} is {1}".format(k, to_string(self[k]))
     #     return envStr
-    
+
 
 def eval(sexp, env=add_globals(Env()), lv=0):
     assert(isa(sexp, SExp))  # Only an object of SExp is evaluable
 
     # evaluating values
-    if isa(sexp, SInt) or isa(sexp, SBool):               
-        return sexp  
+    if isa(sexp, SInt) or isa(sexp, SBool):
+        return sexp
 
     # variable reference
     if isa(sexp, SSymbol):
@@ -93,10 +93,10 @@ def eval(sexp, env=add_globals(Env()), lv=0):
             raise EvalError('lambda requires 2 arguments')
         (_, sparams, sbody) = sexp.children
         e = Echo("lambda", lv)
-        e.ask("lambda creates a function with params %s, and %s as body" % 
+        e.ask("lambda creates a function with params %s, and %s as body" %
                 (html_spanize(sparams.to_lisp_str(), 'parameter'),
                  html_spanize(sbody.to_lisp_str(), 'body')))
-        
+
         # construct the parameter list like: (a b c) and check the type
         params = []
         for i, param in enumerate(sparams.children):
@@ -105,8 +105,8 @@ def eval(sexp, env=add_globals(Env()), lv=0):
             params.append(param.value)
 
         # create a anonymous function with Python's lambda function
-        # lv is passed via the last element of args        
-        retval = lambda *args: eval(sbody, Env(params, args, env), args[len(args)-1] + 1)    
+        # lv is passed via the last element of args
+        retval = lambda *args: eval(sbody, Env(params, args, env), args[len(args)-1] + 1)
         e.answer("function %s created" % str(retval))
         return retval
 
@@ -124,21 +124,21 @@ def eval(sexp, env=add_globals(Env()), lv=0):
             retval = eval(conseq, env, lv+1)
         else:
             e.answer("No, so we do %s" % alt.to_lisp_str())
-            retval = eval(alt, env, lv+1)        
+            retval = eval(alt, env, lv+1)
         e0.answer("It's %s" % retval.to_lisp_str())
         return retval
 
     # (func_name func_args)
-    else: 
+    else:
         e = Echo("func", lv)
         func_name = sexp.children[0]
         func_args = sexp.children[1:]
-        e.ask("call %s with arguments: %s" % 
-                (html_spanize(func_name.to_lisp_str(), 'name'), 
+        e.ask("call %s with arguments: %s" %
+                (html_spanize(func_name.to_lisp_str(), 'name'),
                  ', '.join([arg.to_lisp_str() for arg in func_args])))
         # eval all arguments
-        func_args = [eval(exp, env, lv+1) for exp in sexp.children[1:]]   
-        
+        func_args = [eval(exp, env, lv+1) for exp in sexp.children[1:]]
+
 
         func_name = eval(func_name)
 
@@ -151,5 +151,5 @@ def eval(sexp, env=add_globals(Env()), lv=0):
 
 
 
- 
+
